+++ src/Ficha.jsx
@@ -2,6 +2,7 @@ import { useEffect, useState } from "react";
 import { obtenerDatos, añadirDatos, eliminarDatos} from "./apiUsuarios";
 
 import FormularioDatos from "./FormularioDatos"
+import Datos from "./Datos";
 
 export default function Ficha({nombre, apellido, id_contacto, eliminar}) {
         const [arrDatos, setArrDatos] = useState([]);
@@ -21,14 +22,14 @@ export default function Ficha({nombre, apellido, id_contacto, eliminar}) {
         }
 
         const delDatos = async (id_dato_contacto) => {
-            await eliminarDatos (id_dato_contacto);
-            setArrDatos(arrDatos.filter(d => d.id_dato_contacto != id_dato_contacto));
+            const confirmacion = confirm('¿Confirma que desea eliminar este set de datos?');
+            if (confirmacion) {
+                await eliminarDatos (id_dato_contacto);
+                setArrDatos(arrDatos.filter(d => d.id_dato_contacto != id_dato_contacto));
+        }
         }
 
         var fondo = ''
-        arrDatos.filter(d => d.id_contacto == id_contacto).map((datos) =>(
-            fondo = datos.tipo
-        ))
         
         if (fondo == 'personal') {
             fondo = 'card card-personal'
@@ -39,29 +40,29 @@ export default function Ficha({nombre, apellido, id_contacto, eliminar}) {
         }
 
     return (
-        <div className="card">
+        <div className="card supercard">
             <div className="card-header">
                 {nombre} {apellido}
             </div>
             <div className="card-body">
-            <div className="row row-cols-1 g-4">
                 {
                     arrDatos.filter(datos=> datos.id_contacto == id_contacto)
                     .map((datos) => (
-                        <div className={fondo} key={datos.id_dato_contacto}>
-                            <p>Tipo: {datos.tipo}</p>
-                            <p>Telefono: {datos.telefono}</p>
-                            <p>Correo: {datos.correo}</p>
-                            <p>Dirección : {datos.direccion}</p>
-                            <button onClick={()=>delDatos(datos.id_dato_contacto)} className="btn btn-danger m-1">Eliminar dato de contacto</button>
+                        <div key={datos.id_dato_contacto}>
+                        <Datos
+                                tipo= {datos.tipo}
+                                telefono = {datos.telefono}
+                                correo = {datos.correo}
+                                direccion={datos.direccion}
+                                dato_contacto= {datos.id_dato_contacto}
+                                eliminar_datos = {()=>delDatos(datos.id_dato_contacto)}
+                            ></Datos>
                         </div>
                     ))
                 }
-            </div>
             <FormularioDatos agregarDatos={addDatos} contactoId={id_contacto}></FormularioDatos>
             </div>
-            <button onClick={eliminar} className="btn btn-danger m-1">Eliminar</button>
-            
+            <button onClick={eliminar} className="btn btn-danger m-1">Eliminar contacto</button>
         </div>
     )
 }
